fix(server): dedupe workflow-backed monitor tasks
